requirements.txt
openai==1.82.1
python-dotenv==1.1.0
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
